init: Make scaffold generate mapping based of protocol
