+++ src/commands/init.js
@@ -14,7 +14,8 @@ const { validateStudioNetwork } = require('../command-helpers/studio')
 const { withSpinner, step } = require('../command-helpers/spinner')
 const { fixParameters } = require('../command-helpers/gluegun')
 const { chooseNodeUrl } = require('../command-helpers/node')
-const { abiEvents, generateScaffold, writeScaffold } = require('../old-scaffold')
+const { generateScaffold, writeScaffold } = require('../old-scaffold')
+const { abiEvents } = require('../scaffold/schema')
 const Protocol = require('../protocols')
 // TODO: Use Protocol class to getABI
 const ABI = require('../protocols/ethereum/abi')
