+++ src/scaffold/mapping.js
@@ -1,0 +1,49 @@
+const path = require('path')
+
+const util = require('../codegen/util')
+
+const generateFieldAssignment = path =>
+  `entity.${path.join('_')} = event.params.${path.join('.')}`
+
+const generateFieldAssignments = ({ index, input }) =>
+  input.type === 'tuple'
+    ? util
+        .unrollTuple({ value: input, index, path: [input.name || `param${index}`] })
+        .map(({ path }) => generateFieldAssignment(path))
+    : generateFieldAssignment([input.name || `param${index}`])
+
+const generateEventFieldAssignments = event =>
+  event.inputs.reduce(
+    (acc, input, index) => acc.concat(generateFieldAssignments({ input, index })),
+    [],
+  )
+
+const generateEventIndexingHandlers = (events, contractName) =>
+  `
+  import { ${events.map(
+    event => `${event._alias} as ${event._alias}Event`,
+  )}} from '../generated/${contractName}/${contractName}'
+  import { ${events.map(event => event._alias)} } from '../generated/schema'
+
+  ${events
+    .map(
+      event =>
+        `
+  export function handle${event._alias}(event: ${event._alias}Event): void {
+    let entity = new ${
+      event._alias
+    }(event.transaction.hash.toHex() + '-' + event.logIndex.toString())
+    ${generateEventFieldAssignments(event).join('\n')}
+    entity.save()
+  }
+    `,
+    )
+    .join('\n')}
+`
+
+module.exports = {
+  generateFieldAssignment,
+  generateFieldAssignments,
+  generateEventFieldAssignments,
+  generateEventIndexingHandlers,
+}
